Invert abstractness of generated vs "handwritten" client classes
