+++ java-client/src/main/java/org/opensearch/client/opensearch/OpenSearchAsyncClient.java
@@ -33,13 +33,13 @@
 package org.opensearch.client.opensearch;
 
 import java.io.IOException;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import javax.annotation.Nullable;
-import org.opensearch.client.ApiClient;
 import org.opensearch.client.opensearch._types.ErrorResponse;
 import org.opensearch.client.opensearch._types.OpenSearchException;
-import org.opensearch.client.opensearch.cat.OpenSearchCatClient;
-import org.opensearch.client.opensearch.cluster.OpenSearchClusterClient;
+import org.opensearch.client.opensearch.cat.OpenSearchCatAsyncClient;
+import org.opensearch.client.opensearch.cluster.OpenSearchClusterAsyncClient;
 import org.opensearch.client.opensearch.core.BulkRequest;
 import org.opensearch.client.opensearch.core.BulkResponse;
 import org.opensearch.client.opensearch.core.ClearScrollRequest;
@@ -119,14 +119,13 @@ import org.opensearch.client.opensearch.core.pit.DeletePitRequest;
 import org.opensearch.client.opensearch.core.pit.DeletePitResponse;
 import org.opensearch.client.opensearch.core.pit.ListAllPitRequest;
 import org.opensearch.client.opensearch.core.pit.ListAllPitResponse;
-import org.opensearch.client.opensearch.features.OpenSearchFeaturesClient;
-import org.opensearch.client.opensearch.generic.OpenSearchGenericClient;
-import org.opensearch.client.opensearch.indices.OpenSearchIndicesClient;
-import org.opensearch.client.opensearch.ingest.OpenSearchIngestClient;
-import org.opensearch.client.opensearch.nodes.OpenSearchNodesClient;
-import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownClient;
-import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotClient;
-import org.opensearch.client.opensearch.tasks.OpenSearchTasksClient;
+import org.opensearch.client.opensearch.features.OpenSearchFeaturesAsyncClient;
+import org.opensearch.client.opensearch.indices.OpenSearchIndicesAsyncClient;
+import org.opensearch.client.opensearch.ingest.OpenSearchIngestAsyncClient;
+import org.opensearch.client.opensearch.nodes.OpenSearchNodesAsyncClient;
+import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownAsyncClient;
+import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotAsyncClient;
+import org.opensearch.client.opensearch.tasks.OpenSearchTasksAsyncClient;
 import org.opensearch.client.transport.JsonEndpoint;
 import org.opensearch.client.transport.OpenSearchTransport;
 import org.opensearch.client.transport.TransportOptions;
@@ -137,51 +136,56 @@ import org.opensearch.client.util.ObjectBuilder;
 /**
  * Client for the namespace.
  */
-public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Self>> extends ApiClient<OpenSearchTransport, Self> {
+public class OpenSearchAsyncClient extends OpenSearchAsyncClientBase<OpenSearchAsyncClient> {
+    public OpenSearchAsyncClient(OpenSearchTransport transport) {
+        super(transport, null);
+    }
 
-    public OpenSearchClientBase(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
+    public OpenSearchAsyncClient(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
         super(transport, transportOptions);
     }
 
+    @Override
+    public OpenSearchAsyncClient withTransportOptions(@Nullable TransportOptions transportOptions) {
+        return new OpenSearchAsyncClient(this.transport, transportOptions);
+    }
+
     // ----- Child clients
-    public OpenSearchGenericClient generic() {
-        return new OpenSearchGenericClient(this.transport, this.transportOptions);
+
+    public OpenSearchCatAsyncClient cat() {
+        return new OpenSearchCatAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchCatClient cat() {
-        return new OpenSearchCatClient(this.transport, this.transportOptions);
+    public OpenSearchClusterAsyncClient cluster() {
+        return new OpenSearchClusterAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchClusterClient cluster() {
-        return new OpenSearchClusterClient(this.transport, this.transportOptions);
+    public OpenSearchFeaturesAsyncClient features() {
+        return new OpenSearchFeaturesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchFeaturesClient features() {
-        return new OpenSearchFeaturesClient(this.transport, this.transportOptions);
+    public OpenSearchIndicesAsyncClient indices() {
+        return new OpenSearchIndicesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIndicesClient indices() {
-        return new OpenSearchIndicesClient(this.transport, this.transportOptions);
+    public OpenSearchIngestAsyncClient ingest() {
+        return new OpenSearchIngestAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIngestClient ingest() {
-        return new OpenSearchIngestClient(this.transport, this.transportOptions);
+    public OpenSearchNodesAsyncClient nodes() {
+        return new OpenSearchNodesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchNodesClient nodes() {
-        return new OpenSearchNodesClient(this.transport, this.transportOptions);
+    public OpenSearchShutdownAsyncClient shutdown() {
+        return new OpenSearchShutdownAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchShutdownClient shutdown() {
-        return new OpenSearchShutdownClient(this.transport, this.transportOptions);
+    public OpenSearchSnapshotAsyncClient snapshot() {
+        return new OpenSearchSnapshotAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchSnapshotClient snapshot() {
-        return new OpenSearchSnapshotClient(this.transport, this.transportOptions);
-    }
-
-    public OpenSearchTasksClient tasks() {
-        return new OpenSearchTasksClient(this.transport, this.transportOptions);
+    public OpenSearchTasksAsyncClient tasks() {
+        return new OpenSearchTasksAsyncClient(this.transport, this.transportOptions);
     }
 
     // ----- Endpoint: bulk
@@ -193,14 +197,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public BulkResponse bulk(BulkRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BulkResponse> bulk(BulkRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<BulkRequest, BulkResponse, ErrorResponse> endpoint = (JsonEndpoint<
             BulkRequest,
             BulkResponse,
             ErrorResponse>) BulkRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -213,7 +217,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final BulkResponse bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException, OpenSearchException {
+    public final CompletableFuture<BulkResponse> bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException,
+        OpenSearchException {
         return bulk(fn.apply(new BulkRequest.Builder()).build());
     }
 
@@ -224,8 +229,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public BulkResponse bulk() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<BulkResponse> bulk() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: clear_scroll
@@ -236,14 +241,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ClearScrollResponse clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ClearScrollResponse> clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ClearScrollRequest, ClearScrollResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ClearScrollRequest,
             ClearScrollResponse,
             ErrorResponse>) ClearScrollRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -255,8 +260,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final ClearScrollResponse clearScroll(Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<ClearScrollResponse> clearScroll(
+        Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn
+    ) throws IOException, OpenSearchException {
         return clearScroll(fn.apply(new ClearScrollRequest.Builder()).build());
     }
 
@@ -266,8 +272,12 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ClearScrollResponse clearScroll() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new ClearScrollRequest.Builder().build(), ClearScrollRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ClearScrollResponse> clearScroll() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            new ClearScrollRequest.Builder().build(),
+            ClearScrollRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: count
@@ -278,14 +288,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public CountResponse count(CountRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<CountResponse> count(CountRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CountRequest, CountResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CountRequest,
             CountResponse,
             ErrorResponse>) CountRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -297,7 +307,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final CountResponse count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
+    public final CompletableFuture<CountResponse> count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
         OpenSearchException {
         return count(fn.apply(new CountRequest.Builder()).build());
     }
@@ -308,8 +318,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public CountResponse count() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<CountResponse> count() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: create
@@ -323,14 +333,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> CreateResponse create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<CreateResponse> create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreateRequest<?>, CreateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreateRequest<?>,
             CreateResponse,
             ErrorResponse>) CreateRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -345,8 +355,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> CreateResponse create(Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn)
-        throws IOException, OpenSearchException {
+    public final <TDocument> CompletableFuture<CreateResponse> create(
+        Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn
+    ) throws IOException, OpenSearchException {
         return create(fn.apply(new CreateRequest.Builder<TDocument>()).build());
     }
 
@@ -359,14 +370,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public CreatePitResponse createPit(CreatePitRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<CreatePitResponse> createPit(CreatePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreatePitRequest, CreatePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreatePitRequest,
             CreatePitResponse,
             ErrorResponse>) CreatePitRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -379,8 +390,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final CreatePitResponse createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<CreatePitResponse> createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn)
+        throws IOException, OpenSearchException {
         return createPit(fn.apply(new CreatePitRequest.Builder()).build());
     }
 
@@ -392,14 +403,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public DeleteResponse delete(DeleteRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteResponse> delete(DeleteRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteRequest, DeleteResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteRequest,
             DeleteResponse,
             ErrorResponse>) DeleteRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -411,8 +422,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final DeleteResponse delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<DeleteResponse> delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn)
+        throws IOException, OpenSearchException {
         return delete(fn.apply(new DeleteRequest.Builder()).build());
     }
 
@@ -424,14 +435,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public DeletePitResponse deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeletePitResponse> deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeletePitRequest, DeletePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeletePitRequest,
             DeletePitResponse,
             ErrorResponse>) DeletePitRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -443,8 +454,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final DeletePitResponse deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<DeletePitResponse> deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn)
+        throws IOException, OpenSearchException {
         return deletePit(fn.apply(new DeletePitRequest.Builder()).build());
     }
 
@@ -456,14 +467,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public DeleteByQueryResponse deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteByQueryResponse> deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRequest, DeleteByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRequest,
             DeleteByQueryResponse,
             ErrorResponse>) DeleteByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -475,8 +486,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final DeleteByQueryResponse deleteByQuery(Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<DeleteByQueryResponse> deleteByQuery(
+        Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn
+    ) throws IOException, OpenSearchException {
         return deleteByQuery(fn.apply(new DeleteByQueryRequest.Builder()).build());
     }
 
@@ -489,15 +501,15 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public DeleteByQueryRethrottleResponse deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request) throws IOException,
-        OpenSearchException {
+    public CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRethrottleRequest, DeleteByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRethrottleRequest,
             DeleteByQueryRethrottleResponse,
             ErrorResponse>) DeleteByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -510,7 +522,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final DeleteByQueryRethrottleResponse deleteByQueryRethrottle(
+    public final CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(
         Function<DeleteByQueryRethrottleRequest.Builder, ObjectBuilder<DeleteByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return deleteByQueryRethrottle(fn.apply(new DeleteByQueryRethrottleRequest.Builder()).build());
@@ -524,14 +536,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public DeleteScriptResponse deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteScriptResponse> deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteScriptRequest, DeleteScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteScriptRequest,
             DeleteScriptResponse,
             ErrorResponse>) DeleteScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -543,8 +555,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final DeleteScriptResponse deleteScript(Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<DeleteScriptResponse> deleteScript(
+        Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn
+    ) throws IOException, OpenSearchException {
         return deleteScript(fn.apply(new DeleteScriptRequest.Builder()).build());
     }
 
@@ -556,14 +569,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public BooleanResponse exists(ExistsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BooleanResponse> exists(ExistsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsRequest,
             BooleanResponse,
             ErrorResponse>) ExistsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -575,8 +588,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final BooleanResponse exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<BooleanResponse> exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn)
+        throws IOException, OpenSearchException {
         return exists(fn.apply(new ExistsRequest.Builder()).build());
     }
 
@@ -588,14 +601,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public BooleanResponse existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BooleanResponse> existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsSourceRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsSourceRequest,
             BooleanResponse,
             ErrorResponse>) ExistsSourceRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -607,8 +620,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final BooleanResponse existsSource(Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<BooleanResponse> existsSource(
+        Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn
+    ) throws IOException, OpenSearchException {
         return existsSource(fn.apply(new ExistsSourceRequest.Builder()).build());
     }
 
@@ -620,8 +634,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> ExplainResponse<TDocument> explain(ExplainRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(ExplainRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExplainRequest, ExplainResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ExplainRequest,
@@ -633,7 +647,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -645,7 +659,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> ExplainResponse<TDocument> explain(
+    public final <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(
         Function<ExplainRequest.Builder, ObjectBuilder<ExplainRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -657,18 +671,16 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
     /**
      * Returns the information about the capabilities of fields among multiple
      * indices.
-     *
-     *
      */
 
-    public FieldCapsResponse fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<FieldCapsResponse> fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<FieldCapsRequest, FieldCapsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             FieldCapsRequest,
             FieldCapsResponse,
             ErrorResponse>) FieldCapsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -681,8 +693,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final FieldCapsResponse fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<FieldCapsResponse> fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn)
+        throws IOException, OpenSearchException {
         return fieldCaps(fn.apply(new FieldCapsRequest.Builder()).build());
     }
 
@@ -693,8 +705,12 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public FieldCapsResponse fieldCaps() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new FieldCapsRequest.Builder().build(), FieldCapsRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<FieldCapsResponse> fieldCaps() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            new FieldCapsRequest.Builder().build(),
+            FieldCapsRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: get
@@ -705,8 +721,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> GetResponse<TDocument> get(GetRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<GetResponse<TDocument>> get(GetRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetRequest, GetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetRequest,
@@ -718,7 +734,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -730,7 +746,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> GetResponse<TDocument> get(
+    public final <TDocument> CompletableFuture<GetResponse<TDocument>> get(
         Function<GetRequest.Builder, ObjectBuilder<GetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -745,14 +761,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public GetScriptResponse getScript(GetScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<GetScriptResponse> getScript(GetScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetScriptRequest, GetScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             GetScriptRequest,
             GetScriptResponse,
             ErrorResponse>) GetScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -764,8 +780,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final GetScriptResponse getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<GetScriptResponse> getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn)
+        throws IOException, OpenSearchException {
         return getScript(fn.apply(new GetScriptRequest.Builder()).build());
     }
 
@@ -776,8 +792,12 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      *
      */
-    public GetScriptContextResponse getScriptContext() throws IOException, OpenSearchException {
-        return this.transport.performRequest(GetScriptContextRequest._INSTANCE, GetScriptContextRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<GetScriptContextResponse> getScriptContext() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            GetScriptContextRequest._INSTANCE,
+            GetScriptContextRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: get_script_languages
@@ -787,8 +807,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      *
      */
-    public GetScriptLanguagesResponse getScriptLanguages() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<GetScriptLanguagesResponse> getScriptLanguages() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             GetScriptLanguagesRequest._INSTANCE,
             GetScriptLanguagesRequest._ENDPOINT,
             this.transportOptions
@@ -803,8 +823,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> GetSourceResponse<TDocument> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetSourceRequest, GetSourceResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetSourceRequest,
@@ -816,7 +836,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -828,7 +848,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> GetSourceResponse<TDocument> getSource(
+    public final <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(
         Function<GetSourceRequest.Builder, ObjectBuilder<GetSourceRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -843,14 +863,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> IndexResponse index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<IndexResponse> index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<IndexRequest<?>, IndexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             IndexRequest<?>,
             IndexResponse,
             ErrorResponse>) IndexRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -862,8 +882,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> IndexResponse index(Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn)
-        throws IOException, OpenSearchException {
+    public final <TDocument> CompletableFuture<IndexResponse> index(
+        Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn
+    ) throws IOException, OpenSearchException {
         return index(fn.apply(new IndexRequest.Builder<TDocument>()).build());
     }
 
@@ -875,8 +896,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ListAllPitResponse listAllPit() throws IOException, OpenSearchException {
-        return this.transport.performRequest(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ListAllPitResponse> listAllPit() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: mget
@@ -887,8 +908,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> MgetResponse<TDocument> mget(MgetRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(MgetRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MgetRequest, MgetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MgetRequest,
@@ -900,7 +921,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -912,7 +933,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> MgetResponse<TDocument> mget(
+    public final <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(
         Function<MgetRequest.Builder, ObjectBuilder<MgetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -927,8 +948,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> MsearchResponse<TDocument> msearch(MsearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(MsearchRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchRequest, MsearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchRequest,
@@ -940,7 +961,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -952,7 +973,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> MsearchResponse<TDocument> msearch(
+    public final <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(
         Function<MsearchRequest.Builder, ObjectBuilder<MsearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -967,8 +988,10 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(MsearchTemplateRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
+        MsearchTemplateRequest request,
+        Class<TDocument> tDocumentClass
+    ) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchTemplateRequest, MsearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchTemplateRequest,
@@ -980,7 +1003,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -992,7 +1015,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(
+    public final <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
         Function<MsearchTemplateRequest.Builder, ObjectBuilder<MsearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1007,14 +1030,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public MtermvectorsResponse mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<MtermvectorsResponse> mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MtermvectorsRequest, MtermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             MtermvectorsRequest,
             MtermvectorsResponse,
             ErrorResponse>) MtermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1026,8 +1049,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final MtermvectorsResponse mtermvectors(Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<MtermvectorsResponse> mtermvectors(
+        Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn
+    ) throws IOException, OpenSearchException {
         return mtermvectors(fn.apply(new MtermvectorsRequest.Builder()).build());
     }
 
@@ -1037,8 +1061,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public MtermvectorsResponse mtermvectors() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<MtermvectorsResponse> mtermvectors() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new MtermvectorsRequest.Builder().build(),
             MtermvectorsRequest._ENDPOINT,
             this.transportOptions
@@ -1052,8 +1076,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      *
      */
-    public BooleanResponse ping() throws IOException, OpenSearchException {
-        return this.transport.performRequest(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<BooleanResponse> ping() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: put_script
@@ -1064,14 +1088,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public PutScriptResponse putScript(PutScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<PutScriptResponse> putScript(PutScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<PutScriptRequest, PutScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             PutScriptRequest,
             PutScriptResponse,
             ErrorResponse>) PutScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1083,8 +1107,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final PutScriptResponse putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<PutScriptResponse> putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn)
+        throws IOException, OpenSearchException {
         return putScript(fn.apply(new PutScriptRequest.Builder()).build());
     }
 
@@ -1097,14 +1121,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public RankEvalResponse rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<RankEvalResponse> rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RankEvalRequest, RankEvalResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RankEvalRequest,
             RankEvalResponse,
             ErrorResponse>) RankEvalRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1117,8 +1141,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final RankEvalResponse rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<RankEvalResponse> rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn)
+        throws IOException, OpenSearchException {
         return rankEval(fn.apply(new RankEvalRequest.Builder()).build());
     }
 
@@ -1132,14 +1156,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ReindexResponse reindex(ReindexRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ReindexResponse> reindex(ReindexRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRequest, ReindexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRequest,
             ReindexResponse,
             ErrorResponse>) ReindexRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1153,8 +1177,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final ReindexResponse reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<ReindexResponse> reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn)
+        throws IOException, OpenSearchException {
         return reindex(fn.apply(new ReindexRequest.Builder()).build());
     }
 
@@ -1166,8 +1190,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ReindexResponse reindex() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ReindexResponse> reindex() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: reindex_rethrottle
@@ -1178,14 +1202,15 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public ReindexRethrottleResponse reindexRethrottle(ReindexRethrottleRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(ReindexRethrottleRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRethrottleRequest, ReindexRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRethrottleRequest,
             ReindexRethrottleResponse,
             ErrorResponse>) ReindexRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1197,7 +1222,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final ReindexRethrottleResponse reindexRethrottle(
+    public final CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(
         Function<ReindexRethrottleRequest.Builder, ObjectBuilder<ReindexRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return reindexRethrottle(fn.apply(new ReindexRethrottleRequest.Builder()).build());
@@ -1211,14 +1236,15 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public RenderSearchTemplateResponse renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RenderSearchTemplateRequest, RenderSearchTemplateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RenderSearchTemplateRequest,
             RenderSearchTemplateResponse,
             ErrorResponse>) RenderSearchTemplateRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1230,7 +1256,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final RenderSearchTemplateResponse renderSearchTemplate(
+    public final CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(
         Function<RenderSearchTemplateRequest.Builder, ObjectBuilder<RenderSearchTemplateRequest>> fn
     ) throws IOException, OpenSearchException {
         return renderSearchTemplate(fn.apply(new RenderSearchTemplateRequest.Builder()).build());
@@ -1242,8 +1268,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public RenderSearchTemplateResponse renderSearchTemplate() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new RenderSearchTemplateRequest.Builder().build(),
             RenderSearchTemplateRequest._ENDPOINT,
             this.transportOptions
@@ -1258,7 +1284,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
+    public <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
         ScriptsPainlessExecuteRequest request,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1273,7 +1299,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tResultClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1285,7 +1311,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
+    public final <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
         Function<ScriptsPainlessExecuteRequest.Builder, ObjectBuilder<ScriptsPainlessExecuteRequest>> fn,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1299,8 +1325,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> ScrollResponse<TDocument> scroll(ScrollRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(ScrollRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ScrollRequest, ScrollResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ScrollRequest,
@@ -1312,7 +1338,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1323,7 +1349,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *            {@link ScrollRequest}
      */
 
-    public final <TDocument> ScrollResponse<TDocument> scroll(
+    public final <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(
         Function<ScrollRequest.Builder, ObjectBuilder<ScrollRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1338,8 +1364,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> SearchResponse<TDocument> search(SearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<SearchResponse<TDocument>> search(SearchRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchRequest, SearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchRequest,
@@ -1351,7 +1377,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1363,7 +1389,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> SearchResponse<TDocument> search(
+    public final <TDocument> CompletableFuture<SearchResponse<TDocument>> search(
         Function<SearchRequest.Builder, ObjectBuilder<SearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1379,14 +1405,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public SearchShardsResponse searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<SearchShardsResponse> searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchShardsRequest, SearchShardsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             SearchShardsRequest,
             SearchShardsResponse,
             ErrorResponse>) SearchShardsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1399,8 +1425,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final SearchShardsResponse searchShards(Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<SearchShardsResponse> searchShards(
+        Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn
+    ) throws IOException, OpenSearchException {
         return searchShards(fn.apply(new SearchShardsRequest.Builder()).build());
     }
 
@@ -1411,8 +1438,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public SearchShardsResponse searchShards() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<SearchShardsResponse> searchShards() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new SearchShardsRequest.Builder().build(),
             SearchShardsRequest._ENDPOINT,
             this.transportOptions
@@ -1427,8 +1454,10 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> SearchTemplateResponse<TDocument> searchTemplate(SearchTemplateRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
+        SearchTemplateRequest request,
+        Class<TDocument> tDocumentClass
+    ) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchTemplateRequest, SearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchTemplateRequest,
@@ -1439,7 +1468,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             "org.opensearch.client:Deserializer:_global.search_template.TDocument",
             getDeserializer(tDocumentClass)
         );
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1451,7 +1481,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> SearchTemplateResponse<TDocument> searchTemplate(
+    public final <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
         Function<SearchTemplateRequest.Builder, ObjectBuilder<SearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1468,14 +1498,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public TermsEnumResponse termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<TermsEnumResponse> termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermsEnumRequest, TermsEnumResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermsEnumRequest,
             TermsEnumResponse,
             ErrorResponse>) TermsEnumRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1489,8 +1519,8 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final TermsEnumResponse termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<TermsEnumResponse> termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn)
+        throws IOException, OpenSearchException {
         return termsEnum(fn.apply(new TermsEnumRequest.Builder()).build());
     }
 
@@ -1503,14 +1533,15 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument> TermvectorsResponse termvectors(TermvectorsRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<TermvectorsResponse> termvectors(TermvectorsRequest<TDocument> request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermvectorsRequest<?>, TermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermvectorsRequest<?>,
             TermvectorsResponse,
             ErrorResponse>) TermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1523,7 +1554,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument> TermvectorsResponse termvectors(
+    public final <TDocument> CompletableFuture<TermvectorsResponse> termvectors(
         Function<TermvectorsRequest.Builder<TDocument>, ObjectBuilder<TermvectorsRequest<TDocument>>> fn
     ) throws IOException, OpenSearchException {
         return termvectors(fn.apply(new TermvectorsRequest.Builder<TDocument>()).build());
@@ -1537,7 +1568,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
+    public <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
         UpdateRequest<TDocument, TPartialDocument> request,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1552,7 +1583,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1564,8 +1595,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
-
+    public final <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
         Function<UpdateRequest.Builder<TDocument, TPartialDocument>, ObjectBuilder<UpdateRequest<TDocument, TPartialDocument>>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1581,14 +1611,14 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public UpdateByQueryResponse updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<UpdateByQueryResponse> updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRequest, UpdateByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRequest,
             UpdateByQueryResponse,
             ErrorResponse>) UpdateByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1601,8 +1631,9 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final UpdateByQueryResponse updateByQuery(Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<UpdateByQueryResponse> updateByQuery(
+        Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn
+    ) throws IOException, OpenSearchException {
         return updateByQuery(fn.apply(new UpdateByQueryRequest.Builder()).build());
     }
 
@@ -1615,15 +1646,15 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public UpdateByQueryRethrottleResponse updateByQueryRethrottle(UpdateByQueryRethrottleRequest request) throws IOException,
-        OpenSearchException {
+    public CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(UpdateByQueryRethrottleRequest request)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRethrottleRequest, UpdateByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRethrottleRequest,
             UpdateByQueryRethrottleResponse,
             ErrorResponse>) UpdateByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1636,7 +1667,7 @@ public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Sel
      *
      */
 
-    public final UpdateByQueryRethrottleResponse updateByQueryRethrottle(
+    public final CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(
         Function<UpdateByQueryRethrottleRequest.Builder, ObjectBuilder<UpdateByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return updateByQueryRethrottle(fn.apply(new UpdateByQueryRethrottleRequest.Builder()).build());
