+++ java-codegen/src/main/java/org/opensearch/client/codegen/model/Types.java
@@ -40,11 +40,11 @@ public final class Types {
     }
 
     public static final class Primitive {
-        public static final Type Boolean = Type.builder().name("boolean").build();
-        public static final Type Int = Type.builder().name("int").build();
-        public static final Type Long = Type.builder().name("long").build();
-        public static final Type Float = Type.builder().name("float").build();
-        public static final Type Double = Type.builder().name("double").build();
+        public static final Type Boolean = Type.builder().withName("boolean").build();
+        public static final Type Int = Type.builder().withName("int").build();
+        public static final Type Long = Type.builder().withName("long").build();
+        public static final Type Float = Type.builder().withName("float").build();
+        public static final Type Double = Type.builder().withName("double").build();
     }
 
     public static final class Java {
@@ -52,57 +52,57 @@ public final class Types {
 
         public static final class Io {
             public static final String PACKAGE = Java.PACKAGE + ".io";
-            public static final Type IOException = Type.builder().pkg(PACKAGE).name("IOException").build();
+            public static final Type IOException = Type.builder().withPackage(PACKAGE).withName("IOException").build();
         }
 
         public static final class Lang {
             public static final String PACKAGE = Java.PACKAGE + ".lang";
-            public static final Type String = Type.builder().pkg(PACKAGE).name("String").build();
-            public static final Type Character = Type.builder().pkg(PACKAGE).name("Character").build();
-            public static final Type Boolean = Type.builder().pkg(PACKAGE).name("Boolean").build();
-            public static final Type Byte = Type.builder().pkg(PACKAGE).name("Byte").build();
-            public static final Type Short = Type.builder().pkg(PACKAGE).name("Short").build();
-            public static final Type Integer = Type.builder().pkg(PACKAGE).name("Integer").build();
-            public static final Type Long = Type.builder().pkg(PACKAGE).name("Long").build();
-            public static final Type Float = Type.builder().pkg(PACKAGE).name("Float").build();
-            public static final Type Double = Type.builder().pkg(PACKAGE).name("Double").build();
-            public static final Type Object = Type.builder().pkg(PACKAGE).name("Object").build();
+            public static final Type String = Type.builder().withPackage(PACKAGE).withName("String").build();
+            public static final Type Character = Type.builder().withPackage(PACKAGE).withName("Character").build();
+            public static final Type Boolean = Type.builder().withPackage(PACKAGE).withName("Boolean").build();
+            public static final Type Byte = Type.builder().withPackage(PACKAGE).withName("Byte").build();
+            public static final Type Short = Type.builder().withPackage(PACKAGE).withName("Short").build();
+            public static final Type Integer = Type.builder().withPackage(PACKAGE).withName("Integer").build();
+            public static final Type Long = Type.builder().withPackage(PACKAGE).withName("Long").build();
+            public static final Type Float = Type.builder().withPackage(PACKAGE).withName("Float").build();
+            public static final Type Double = Type.builder().withPackage(PACKAGE).withName("Double").build();
+            public static final Type Object = Type.builder().withPackage(PACKAGE).withName("Object").build();
         }
 
         public static final class Util {
             public static final String PACKAGE = Java.PACKAGE + ".util";
-            public static final Type HashMap = Type.builder().pkg(PACKAGE).name("HashMap").build();
+            public static final Type HashMap = Type.builder().withPackage(PACKAGE).withName("HashMap").build();
 
             public static Type Map(Type keyType, Type valueType) {
                 return Map.withTypeParams(keyType, valueType);
             }
 
-            public static final Type Map = Type.builder().pkg(PACKAGE).name("Map").build();
+            public static final Type Map = Type.builder().withPackage(PACKAGE).withName("Map").build();
 
             public static Type MapEntry(Type keyType, Type valueType) {
-                return Type.builder().pkg(PACKAGE).name("Map.Entry").typeParams(keyType, valueType).build();
+                return Type.builder().withPackage(PACKAGE).withName("Map.Entry").withTypeParameters(keyType, valueType).build();
             }
 
             public static Type List(Type valueType) {
-                return Type.builder().pkg(PACKAGE).name("List").typeParams(valueType).build();
+                return Type.builder().withPackage(PACKAGE).withName("List").withTypeParameters(valueType).build();
             }
 
             public static final class Concurrent {
                 public static final String PACKAGE = Util.PACKAGE + ".concurrent";
-                public static final Type CompletableFuture = Type.builder().pkg(PACKAGE).name("CompletableFuture").build();
+                public static final Type CompletableFuture = Type.builder().withPackage(PACKAGE).withName("CompletableFuture").build();
             }
 
             public static final class Function {
                 public static final String PACKAGE = Util.PACKAGE + ".function";
 
                 public static Type Function(Type argType, Type returnType) {
-                    return Type.builder().pkg(PACKAGE).name("Function").typeParams(argType, returnType).build();
+                    return Type.builder().withPackage(PACKAGE).withName("Function").withTypeParameters(argType, returnType).build();
                 }
             }
 
             public static final class Stream {
                 public static final String PACKAGE = Util.PACKAGE + ".stream";
-                public static final Type Collectors = Type.builder().pkg(PACKAGE).name("Collectors").build();
+                public static final Type Collectors = Type.builder().withPackage(PACKAGE).withName("Collectors").build();
             }
         }
     }
@@ -112,8 +112,8 @@ public final class Types {
 
         public static final class Annotation {
             public static final String PACKAGE = Javax.PACKAGE + ".annotation";
-            public static final Type Generated = Type.builder().pkg(PACKAGE).name("Generated").build();
-            public static final Type Nullable = Type.builder().pkg(PACKAGE).name("Nullable").build();
+            public static final Type Generated = Type.builder().withPackage(PACKAGE).withName("Generated").build();
+            public static final Type Nullable = Type.builder().withPackage(PACKAGE).withName("Nullable").build();
         }
     }
 
@@ -124,20 +124,23 @@ public final class Types {
             return ApiClient.withTypeParams(transport, client);
         }
 
-        public static final Type ApiClient = Type.builder().pkg(PACKAGE).name("ApiClient").build();
+        public static final Type ApiClient = Type.builder().withPackage(PACKAGE).withName("ApiClient").build();
 
         public static final class Json {
             public static final String PACKAGE = Client.PACKAGE + ".json";
-            public static final Type JsonData = Type.builder().pkg(PACKAGE).name("JsonData").build();
-            public static final Type JsonpDeserializable = Type.builder().pkg(PACKAGE).name("JsonpDeserializable").build();
-            public static final Type JsonpDeserializer = Type.builder().pkg(PACKAGE).name("JsonpDeserializer").build();
-            public static final Type JsonEnum = Type.builder().pkg(PACKAGE).name("JsonEnum").build();
-            public static final Type JsonpMapper = Type.builder().pkg(PACKAGE).name("JsonpMapper").build();
-            public static final Type JsonpSerializable = Type.builder().pkg(PACKAGE).name("JsonpSerializable").build();
-            public static final Type ObjectBuilderDeserializer = Type.builder().pkg(PACKAGE).name("ObjectBuilderDeserializer").build();
-            public static final Type ObjectDeserializer = Type.builder().pkg(PACKAGE).name("ObjectDeserializer").build();
-            public static final Type PlainJsonSerializable = Type.builder().pkg(PACKAGE).name("PlainJsonSerializable").build();
-            public static final Type UnionDeserializer = Type.builder().pkg(PACKAGE).name("UnionDeserializer").build();
+            public static final Type JsonData = Type.builder().withPackage(PACKAGE).withName("JsonData").build();
+            public static final Type JsonpDeserializable = Type.builder().withPackage(PACKAGE).withName("JsonpDeserializable").build();
+            public static final Type JsonpDeserializer = Type.builder().withPackage(PACKAGE).withName("JsonpDeserializer").build();
+            public static final Type JsonEnum = Type.builder().withPackage(PACKAGE).withName("JsonEnum").build();
+            public static final Type JsonpMapper = Type.builder().withPackage(PACKAGE).withName("JsonpMapper").build();
+            public static final Type JsonpSerializable = Type.builder().withPackage(PACKAGE).withName("JsonpSerializable").build();
+            public static final Type ObjectBuilderDeserializer = Type.builder()
+                .withPackage(PACKAGE)
+                .withName("ObjectBuilderDeserializer")
+                .build();
+            public static final Type ObjectDeserializer = Type.builder().withPackage(PACKAGE).withName("ObjectDeserializer").build();
+            public static final Type PlainJsonSerializable = Type.builder().withPackage(PACKAGE).withName("PlainJsonSerializable").build();
+            public static final Type UnionDeserializer = Type.builder().withPackage(PACKAGE).withName("UnionDeserializer").build();
         }
 
         public static final class OpenSearch {
@@ -145,48 +148,48 @@ public final class Types {
 
             public static final class _Types {
                 public static final String PACKAGE = OpenSearch.PACKAGE + "._types";
-                public static final Type ErrorResponse = Type.builder().pkg(PACKAGE).name("ErrorResponse").build();
-                public static final Type OpenSearchException = Type.builder().pkg(PACKAGE).name("OpenSearchException").build();
-                public static final Type RequestBase = Type.builder().pkg(PACKAGE).name("RequestBase").build();
-                public static final Type Time = Type.builder().pkg(PACKAGE).name("Time").build();
+                public static final Type ErrorResponse = Type.builder().withPackage(PACKAGE).withName("ErrorResponse").build();
+                public static final Type OpenSearchException = Type.builder().withPackage(PACKAGE).withName("OpenSearchException").build();
+                public static final Type RequestBase = Type.builder().withPackage(PACKAGE).withName("RequestBase").build();
+                public static final Type Time = Type.builder().withPackage(PACKAGE).withName("Time").build();
             }
         }
 
         public static final class Transport {
             public static final String PACKAGE = Client.PACKAGE + ".transport";
-            public static final Type Endpoint = Type.builder().pkg(PACKAGE).name("Endpoint").build();
+            public static final Type Endpoint = Type.builder().withPackage(PACKAGE).withName("Endpoint").build();
 
             public static Type JsonEndpoint(Type requestType, Type responseType, Type errorType) {
                 return JsonEndpoint.withTypeParams(requestType, responseType, errorType);
             }
 
-            public static final Type JsonEndpoint = Type.builder().pkg(PACKAGE).name("JsonEndpoint").build();
-            public static final Type OpenSearchTransport = Type.builder().pkg(PACKAGE).name("OpenSearchTransport").build();
-            public static final Type TransportOptions = Type.builder().pkg(PACKAGE).name("TransportOptions").build();
+            public static final Type JsonEndpoint = Type.builder().withPackage(PACKAGE).withName("JsonEndpoint").build();
+            public static final Type OpenSearchTransport = Type.builder().withPackage(PACKAGE).withName("OpenSearchTransport").build();
+            public static final Type TransportOptions = Type.builder().withPackage(PACKAGE).withName("TransportOptions").build();
 
             public static final class Endpoints {
                 public static final String PACKAGE = Transport.PACKAGE + ".endpoints";
-                public static final Type SimpleEndpoint = Type.builder().pkg(PACKAGE).name("SimpleEndpoint").build();
+                public static final Type SimpleEndpoint = Type.builder().withPackage(PACKAGE).withName("SimpleEndpoint").build();
             }
         }
 
         public static final class Util {
             public static final String PACKAGE = Client.PACKAGE + ".util";
-            public static final Type ApiTypeHelper = Type.builder().pkg(PACKAGE).name("ApiTypeHelper").build();
+            public static final Type ApiTypeHelper = Type.builder().withPackage(PACKAGE).withName("ApiTypeHelper").build();
 
             public static Type ObjectBuilder(Type type) {
                 return ObjectBuilder.withTypeParams(type);
             }
 
-            public static final Type ObjectBuilder = Type.builder().pkg(PACKAGE).name("ObjectBuilder").build();
-            public static final Type ObjectBuilderBase = Type.builder().pkg(PACKAGE).name("ObjectBuilderBase").build();
+            public static final Type ObjectBuilder = Type.builder().withPackage(PACKAGE).withName("ObjectBuilder").build();
+            public static final Type ObjectBuilderBase = Type.builder().withPackage(PACKAGE).withName("ObjectBuilderBase").build();
 
             public static Type TaggedUnion(Type tagType, Type baseType) {
                 return TaggedUnion.withTypeParams(tagType, baseType);
             }
 
-            public static final Type TaggedUnion = Type.builder().pkg(PACKAGE).name("TaggedUnion").build();
-            public static final Type TaggedUnionUtils = Type.builder().pkg(PACKAGE).name("TaggedUnionUtils").build();
+            public static final Type TaggedUnion = Type.builder().withPackage(PACKAGE).withName("TaggedUnion").build();
+            public static final Type TaggedUnionUtils = Type.builder().withPackage(PACKAGE).withName("TaggedUnionUtils").build();
         }
     }
 
@@ -198,7 +201,7 @@ public final class Types {
 
             public static final class Stream {
                 public static final String PACKAGE = Json.PACKAGE + ".stream";
-                public static final Type JsonGenerator = Type.builder().pkg(PACKAGE).name("JsonGenerator").build();
+                public static final Type JsonGenerator = Type.builder().withPackage(PACKAGE).withName("JsonGenerator").build();
             }
         }
     }
