+++ java-codegen/src/main/java/org/opensearch/client/codegen/openapi/OpenApiSchema.java
@@ -21,8 +21,10 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.jetbrains.annotations.NotNull;
 import org.opensearch.client.codegen.utils.Lists;
 import org.opensearch.client.codegen.utils.Maps;
+import org.opensearch.client.codegen.utils.ObjectBuilderBase;
 import org.opensearch.client.codegen.utils.Sets;
 import org.opensearch.client.codegen.utils.Versions;
 import org.semver4j.Semver;
@@ -311,7 +313,7 @@ public class OpenApiSchema extends OpenApiRefElement<OpenApiSchema> {
         return new Builder();
     }
 
-    public static class Builder {
+    public static class Builder extends ObjectBuilderBase<OpenApiSchema, Builder> {
         @Nullable
         private OpenApiElement<?> parent;
         @Nullable
@@ -353,6 +355,15 @@ public class OpenApiSchema extends OpenApiRefElement<OpenApiSchema> {
         @Nullable
         private Semver versionRemoved;
 
+        private Builder() {
+            super(OpenApiSchema::new);
+        }
+
+        @Override
+        protected @NotNull Builder self() {
+            return this;
+        }
+
         @Nonnull
         public Builder withPointer(@Nonnull JsonPointer pointer) {
             this.pointer = Objects.requireNonNull(pointer, "pointer must not be null");
@@ -379,11 +390,6 @@ public class OpenApiSchema extends OpenApiRefElement<OpenApiSchema> {
         public Builder withAllOf(@Nullable List<OpenApiSchema> allOf) {
             this.allOf = allOf;
             return this;
-        }
-
-        @Nonnull
-        public OpenApiSchema build() {
-            return new OpenApiSchema(this);
         }
     }
 }
