+++ java-codegen/src/main/java/org/opensearch/client/codegen/renderer/JavaCodeFormatter.java
@@ -25,6 +25,7 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 import org.opensearch.client.codegen.exceptions.JavaFormatterException;
 import org.opensearch.client.codegen.utils.MavenArtifactResolver;
+import org.opensearch.client.codegen.utils.ObjectBuilderBase;
 
 public class JavaCodeFormatter implements AutoCloseable {
     private final Formatter formatter;
@@ -75,9 +76,18 @@ public class JavaCodeFormatter implements AutoCloseable {
         return new Builder();
     }
 
-    public static final class Builder {
+    public static final class Builder extends ObjectBuilderBase<JavaCodeFormatter, Builder> {
         private Path rootDir;
         private File eclipseFormatterConfig;
+
+        private Builder() {
+            super(JavaCodeFormatter::new);
+        }
+
+        @Override
+        protected @Nonnull Builder self() {
+            return this;
+        }
 
         @Nonnull
         public Builder withRootDir(@Nonnull Path rootDir) {
@@ -89,11 +99,6 @@ public class JavaCodeFormatter implements AutoCloseable {
         public Builder withEclipseFormatterConfig(@Nonnull File eclipseFormatterConfig) {
             this.eclipseFormatterConfig = Objects.requireNonNull(eclipseFormatterConfig, "eclipseFormatterConfig must not be null");
             return this;
-        }
-
-        @Nonnull
-        public JavaCodeFormatter build() {
-            return new JavaCodeFormatter(this);
         }
     }
 }
