+++ java-codegen/src/main/java/org/opensearch/client/codegen/utils/ObjectBuilder.java
@@ -1,0 +1,16 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.client.codegen.utils;
+
+import javax.annotation.Nonnull;
+
+public interface ObjectBuilder<T> {
+    @Nonnull
+    T build();
+}
